% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoBand_Hg38.R
\docType{data}
\name{cytoBand_Hg38}
\alias{cytoBand_Hg38}
\title{Cytogenetics data for Hg38 genome}
\format{
An object of class \code{"data frame"}
}
\source{
\href{https://genome.ucsc.edu/}{UCSC Genome Browser}
}
\usage{
data(cytoBand_Hg38)
}
\description{
Cytogenetics data for Hg38 genome
}
\examples{
data(cytoBand_Hg38)
head(cytoBand_Hg38)

}
\references{
Karolchik, Donna, et al. Nucleic acids research 31.1 (2003): 51-54.
(\href{https://pubmed.ncbi.nlm.nih.gov/33221922/}{PubMed})
}
\keyword{datasets}
