% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hg38_gtf.R
\docType{data}
\name{Hg38_gtf}
\alias{Hg38_gtf}
\title{GTF data for Hg38 genome}
\format{
An object of class \code{"data frame"}
}
\source{
\href{https://www.gencodegenes.org/human/}{Gencode Archive}
}
\usage{
data(Hg38_gtf)
}
\value{
Gene location data for Hg38 genome
}
\description{
Gene location data for Hg38 genome
}
\examples{
data(Hg38_gtf)
head(Hg38_gtf)

}
\references{
Frankish, Adam, et al. Nucleic acids research 47.D1 (2019): D766-D773.
(\href{https://pubmed.ncbi.nlm.nih.gov/30357393/}{PubMed})
}
\keyword{datasets}
