% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData.R
\docType{data}
\name{SimData}
\alias{SimData}
\title{Simulation data to examplify the usage of the method}
\format{
An object of class \code{"data frame"}
}
\usage{
data(SimData)
}
\value{
Simulation data to examplify the usage of the method
}
\description{
Simulation data to examplify the usage of the method
}
\examples{
data(SimData)
dim(SimData)

}
\keyword{datasets}
