% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimDataSce.R
\docType{data}
\name{SimDataSce}
\alias{SimDataSce}
\title{Simulation SingleCellExperiment object to examplify the usage of the method}
\format{
A SingleCellExperiment object
}
\usage{
data(SimDataSce)
}
\value{
Simulation SingleCellExperiment object to examplify the usage of the method
}
\description{
Simulation SingleCellExperiment object to examplify the usage of the method
}
\examples{
data(SimDataSce)

}
\keyword{datasets}
